+++ src/classes/index.ts
@@ -1,6 +1,8 @@
+export * from './async-fifo-queue';
 export * from './backoffs';
 export * from './child-pool';
 export * from './child-processor';
+export * from './delayed-error';
 export * from './flow-producer';
 export * from './job';
 // export * from './master'; this file must not be exported
@@ -15,4 +17,5 @@ export * from './repeat';
 export * from './sandbox';
 export * from './scripts';
 export * from './unrecoverable-error';
+export * from './waiting-children-error';
 export * from './worker';